+++ HISTORY.rst
@@ -3,6 +3,12 @@
 Release History
 ---------------
 
+0.2.0 (2014-07-28)
+~~~~~~~~~~~~~~~~~~
+
+* Add template method to fetch and extract content from a path (URL or file)
+* Reorganize code into separate files
+
 0.1.0 (2014-07-23)
 ~~~~~~~~~~~~~~~~~~